SES-13
Refactoring of Network module:
1. now IChannel and ITerminal are templates
2. old IChannel and ITerminal has been renamed to IBinaryChannel and
IBinaryTerminal
3. sessionId type changed from size_t to uint32_t
4. network::MessagePtr replaced with network::BinaryMessage

+++ Network/BufferedProtobufTerminal.cpp
@@ -3,7 +3,7 @@
 namespace network {
 
 void BufferedProtobufTerminal::onMessageReceived(
-    size_t nSessionId, spex::ICommutator &&message)
+    uint32_t nSessionId, spex::ICommutator&& message)
 {
   m_messages.emplace_back(nSessionId, std::move(message));
 }
